+++ src/black/parsing.py
@@ -5,7 +5,7 @@ Parse Python code and perform AST validation.
 import ast
 import sys
 import warnings
-from typing import Iterable, Iterator, List, Set, Tuple
+from typing import Collection, Iterator, List, Set, Tuple
 
 from black.mode import VERSION_TO_FEATURES, Feature, TargetVersion, supports_feature
 from black.nodes import syms
@@ -52,21 +52,19 @@ def get_grammars(target_versions: Set[TargetVersion]) -> List[Grammar]:
     return grammars
 
 
-def lib2to3_parse(src_txt: str, target_versions: Iterable[TargetVersion] = ()) -> Node:
+def lib2to3_parse(
+    src_txt: str, target_versions: Collection[TargetVersion] = ()
+) -> Node:
     """Given a string with source, return the lib2to3 Node."""
     if not src_txt.endswith("\n"):
         src_txt += "\n"
 
     grammars = get_grammars(set(target_versions))
-
-    version_names = [version.name for version in target_versions]
-    grammar_names = [
-        f"{grammar.version[0]}.{grammar.version[1]}" for grammar in grammars
-    ]
-    error_string = (
-        f"The assigned grammars ({', '.join(sorted(grammar_names))}) for your Python"
-        f" versions ({', '.join(sorted(version_names))}) all failed to parse"
-    )
+    if target_versions:
+        max_tv = max(target_versions, key=lambda tv: tv.value)
+        tv_str = f" for target version {max_tv.pretty()}"
+    else:
+        tv_str = ""
 
     errors = {}
     for grammar in grammars:
@@ -83,14 +81,14 @@ def lib2to3_parse(src_txt: str, target_versions: Iterable[TargetVersion] = ()) -
             except IndexError:
                 faulty_line = "<line number missing in source>"
             errors[grammar.version] = InvalidInput(
-                f"Cannot parse: {error_string} line: {lineno}:{column}: {faulty_line}"
+                f"Cannot parse{tv_str}: {lineno}:{column}: {faulty_line}"
             )
 
         except TokenError as te:
             # In edge cases these are raised; and typically don't have a "faulty_line".
             lineno, column = te.args[1]
             errors[grammar.version] = InvalidInput(
-                f"Cannot parse: {error_string}: {lineno}:{column}: {te.args[0]}"
+                f"Cannot parse{tv_str}: {lineno}:{column}: {te.args[0]}"
             )
 
     else: